refactor(util/Platform): unify with SystemInfo class
